cleaned up auto update; removed refresh button; limited capability to only running one optimization at a time
