+++ electron/ui/src/services/sidebar.service.js
@@ -1,12 +1,3 @@
-
-
-export const fetchScenarios = () => {
-    return fetch('http://localhost:8001/get_scenario_list/', {
-        method: 'GET', 
-        mode: 'cors'
-    });
-}; 
-
 export const uploadExcelSheet = (data) => {
     return fetch('http://localhost:8001/upload', {
         method: 'POST',
